Comments model - doc

+++ src/models/Comments.php
@@ -1,5 +1,8 @@
 <?php
+declare (strict_types=1);
+
 namespace models;
+
 
 use core\Model;
 use models\obj\Message;
@@ -7,11 +10,11 @@ use models\obj\Comment;
 
 
 /**
- * Responsible for managing comments table.
+ * Responsible for managing 'comments' table.
  * 
  * @author		William Niemiec &lt; williamniemiec@hotmail.com &gt;
- * @version		1.0
- * @since		1.0
+ * @version		1.0.0
+ * @since		1.0.0
  */
 class Comments extends Model
 {
@@ -19,7 +22,7 @@ class Comments extends Model
     //        Constructor
     //-------------------------------------------------------------------------
     /**
-     * Creates doubts table manager.
+     * Creates 'comments' table manager.
      *
      * @apiNote     It will connect to the database when it is instantiated
      */
@@ -36,54 +39,74 @@ class Comments extends Model
      * Adds a new comment to a class.
      * 
      * @param       int $id_student Student id
-     * @param       int $id_class Class id
-     * @param       string $text Doubt content
+     * @param       int $id_module Module id that the class belongs
+     * @param       int $class_order Class order within the module
+     * @param       string $text Comment content
      * 
-     * @return      boolean If the doubt was successfully added
+     * @return      bool If the comment was successfully added
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function newComment($id_student, $id_module, $class_order, $text)
+    public function newComment(int $id_student, int $id_module, int $class_order, string $text) : bool
     {
-//         if (empty($id_class) && $id_class <= 0 || empty($id_student) && $id_student <= 0)
-//             return false;
+        if (empty($id_student) || $id_student <= 0)
+            throw new \InvalidArgumentException("Invalid id_student");
         
+        if (empty($id_module) || $id_module <= 0)
+            throw new \InvalidArgumentException("Invalid id_module");
         
-        //$students = new Students();
+        if (empty($class_order) || $class_order <= 0)
+            throw new \InvalidArgumentException("Invalid class_order");
         
-//         if (!$classes->exist($id_class) || !$students->exist($id_student)) { return false; }
+        if (empty($text))
+            throw new \InvalidArgumentException("Invalid text");
         
+        // Query construction
         $sql = $this->db->prepare("
             INSERT INTO comments 
             (id_student, id_module, class_order, date, text) 
             VALUES (?, ?, ?, NOW(), ?)
         ");
         
+        // Executes query
         $sql->execute(array($id_student, $id_module, $class_order, $text));
         
-        return $sql->rowCount() > 0;
+        return $sql && $sql->rowCount() > 0;
     }
     
     /**
      * Gets comments from a class.
      * 
-     * @param       int $id_class Class id
+     * @param       int $id_module Module id that the class belongs
+     * @param       int $class_order Class order within the module
      * 
-     * @return      array Doubts from this class
+     * @return      Comment[] Comments from this class
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function getComments($id_module, $class_order)
-    {//
-       // if (empty($id_class) && $id_class <= 0) { return array(); }
+    public function getComments(int $id_module, int $class_order) : array
+    {
+        if (empty($id_module) || $id_module <= 0)
+            throw new \InvalidArgumentException("Invalid id_module");
+            
+        if (empty($class_order) || $class_order <= 0)
+            throw new \InvalidArgumentException("Invalid class_order");
         
         $response = array();
         
+        // Query construction
         $sql = $this->db->prepare("
             SELECT  student_name, student_photo, text, date
-            FROM    comments NATURAL JOIN students
+            FROM    comments
+                    NATURAL JOIN students
             WHERE   id_module = ? AND class_order = ?
         ");
         
+        // Executes query
         $sql->execute(array($id_module, $class_order));
         
-        if ($sql->rowCount() > 0) {
+        // Parses results
+        if ($sql && $sql->rowCount() > 0) {
             foreach ($sql->fetchAll(\PDO::FETCH_ASSOC) as $comment) {
                 $response[] = new Comment(
                     $comment['id_comment'], 
@@ -100,29 +123,39 @@ class Comments extends Model
     /**
      * Replies a comment.
      * 
-     * @param       int $id_comment Doubt id
-     * @param       int $id_student Student id that will reply the doubt
+     * @param       int $id_comment Comment id
+     * @param       int $id_student Student id that replies the comment
      * @param       string $text Reply content
      * 
-     * @return      boolean If reply was sucessfully added
+     * @return      int Reply id added or -1 if reply was not added
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function newReply($id_comment, $id_student, $text) 
+    public function newReply(int $id_comment, int $id_student, string $text) : int 
     {
-        if (empty($id_comment) || $id_comment <= 0) { return -1; }
-        if (empty($id_student) || $id_student <= 0)   { return -1; }
-        if (empty($text))                       { return false; }
+        if (empty($id_comment) || $id_comment <= 0)
+            throw new \InvalidArgumentException("Invalid id_comment");
+            
+        if (empty($id_student) || $id_student <= 0)
+            throw new \InvalidArgumentException("Invalid id_student");
+        
+        if (empty($text))
+            throw new \InvalidArgumentException("Invalid text");
         
         $response = -1;
         
+        // Query construction
         $sql = $this->db->prepare("
             INSERT INTO comment_replies 
             (id_comment, id_student, date, text) 
             VALUES (?, ?, NOW(), ?)
         ");
         
+        // Executes query
         $sql->execute(array($id_comment, $id_student, $text));
         
-        if ($sql->rowCount() > 0) {
+        // Parses results
+        if ($sql && $sql->rowCount() > 0) {
             $response = $this->db->lastInsertId();
         }
         
@@ -132,25 +165,31 @@ class Comments extends Model
     /**
      * Gets replies from a comment.
      * 
-     * @param       int $id_comment Doubt id
+     * @param       int $id_comment Comment id
      * 
-     * @return      array Replies from this doubt
+     * @return      array Replies from this comment
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function getReplies($id_comment)
+    public function getReplies(int $id_comment) : array
     {
-//         if (empty($id_comment) || $id_comment <= 0) { return array(); }
+        if (empty($id_comment) || $id_comment <= 0)
+            throw new \InvalidArgumentException("Invalid id_comment");
         
         $response = array();
         
+        // Query construction
         $sql = $this->db->prepare("
             SELECT  student_name, student_photo, text, date
             FROM    comment_replies NATURAL JOIN students
             WHERE   id_comment = ?
         ");
         
+        // Executes query
         $sql->execute(array($id_comment));
         
-        if ($sql->rowCount() > 0) {
+        // Parses results
+        if ($sql && $sql->rowCount() > 0) {
             $replies = $sql->fetchAll(\PDO::FETCH_ASSOC);
             $students = new Students();
             
@@ -172,18 +211,24 @@ class Comments extends Model
      * @param       int $id_comment Doubt id
      * 
      * @return      boolean If doubt was sucessfully deleted
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function deleteComment($id_comment)
+    public function deleteComment(int $id_comment) : bool
     {
-        if (empty($id_comment) && $id_comment <= 0) { return false; }
+        if (empty($id_comment) || $id_comment <= 0)
+            throw new \InvalidArgumentException("Invalid id_comment");
         
+        // Query construction
         $sql = $this->db->prepare("
-            DELETE FROM comments WHERE id_comment = ?
+            DELETE FROM comments 
+            WHERE id_comment = ?
         ");
         
+        // Executes query
         $sql->execute(array($id_comment));
         
-        return $sql->rowCount() > 0;
+        return $sql && $sql->rowCount() > 0;
     }
     
     /**
@@ -192,17 +237,26 @@ class Comments extends Model
      * @param       int $id_reply Reply id
      * 
      * @return      boolean If reply was sucessfully deleted
+     * 
+     * @throws      \InvalidArgumentException If any argument is invalid 
      */
-    public function deleteReply($id_comment, $id_student)
+    public function deleteReply(int $id_comment, int $id_student) : bool
     {
-//         if (empty($id_reply) && $id_reply <= 0) { return false; }
+        if (empty($id_comment) || $id_comment <= 0)
+            throw new \InvalidArgumentException("Invalid id_comment");
+            
+        if (empty($id_student) || $id_student <= 0)
+            throw new \InvalidArgumentException("Invalid id_student");
         
+        // Query construction
         $sql = $this->db->prepare("
             DELETE FROM comment_replies 
             WHERE id_comment = ? AND id_student = ? 
         ");
+        
+        // Executes query
         $sql->execute(array($id_comment, $id_student));
         
-        return $sql->rowCount() > 0;
+        return $sql && $sql->rowCount() > 0;
     }
 }
